add createPureReaction and try to rewrite lazy memo
